requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.30.0
pillow>=8.2.0
numpy>=1.19.0
matplotlib>=3.3.0
pandas>=1.2.0
tqdm>=4.60.0